Use text for shortDesc
